+++ axelor-helpdesk/src/main/java/com/axelor/apps/helpdesk/service/TimerTicketServiceImpl.java
@@ -1,4 +1,4 @@
-package com.axelor.apps.project.service;
+package com.axelor.apps.helpdesk.service;
 
 import com.axelor.apps.base.db.Timer;
 import com.axelor.apps.base.db.TimerHistory;
@@ -6,19 +6,19 @@ import com.axelor.apps.base.db.repo.TimerHistoryRepository;
 import com.axelor.apps.base.db.repo.TimerRepository;
 import com.axelor.apps.base.service.timer.AbstractTimerService;
 import com.axelor.apps.base.service.user.UserService;
+import com.axelor.apps.helpdesk.db.Ticket;
 import com.axelor.auth.db.User;
 import com.axelor.db.Model;
 import com.axelor.exception.AxelorException;
-import com.axelor.team.db.TeamTask;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
 import java.time.Duration;
 import java.time.LocalDateTime;
 
-public class TimerTeamTaskServiceImpl extends AbstractTimerService implements TimerTeamTaskService {
+public class TimerTicketServiceImpl extends AbstractTimerService implements TimerTicketService {
 
   @Inject
-  public TimerTeamTaskServiceImpl(
+  public TimerTicketServiceImpl(
       TimerRepository timerRepository,
       TimerHistoryRepository timerHistoryRepository,
       UserService userService) {
@@ -28,9 +28,10 @@ public class TimerTeamTaskServiceImpl extends AbstractTimerService implements Ti
   @Override
   public Timer find(Model model) {
     User user = userService.getUser();
-    TeamTask task = (TeamTask) model;
+    Ticket ticket = (Ticket) model;
 
-    return task.getTimerList()
+    return ticket
+        .getTimerList()
         .stream()
         .filter(t -> t.getAssignedTo() == user)
         .findFirst()
@@ -38,15 +39,15 @@ public class TimerTeamTaskServiceImpl extends AbstractTimerService implements Ti
   }
 
   @Override
-  @Transactional(rollbackOn = {Exception.class, RuntimeException.class})
+  @Transactional
   public TimerHistory start(Model model, Timer timer, LocalDateTime dateTime)
       throws AxelorException {
-    TeamTask task = (TeamTask) model;
+    Ticket ticket = (Ticket) model;
 
     boolean isNewTimer = timer == null;
     timer = tryStartOrCreate(timer);
     if (isNewTimer) {
-      task.addTimerListItem(timer);
+      ticket.addTimerListItem(timer);
     }
 
     TimerHistory history = new TimerHistory();
@@ -58,30 +59,30 @@ public class TimerTeamTaskServiceImpl extends AbstractTimerService implements Ti
   }
 
   @Override
-  public Timer find(TeamTask task) {
-    return find((Model) task);
+  public Timer find(Ticket ticket) {
+    return find((Model) ticket);
   }
 
   @Override
-  public TimerHistory start(TeamTask task, LocalDateTime dateTime) throws AxelorException {
+  public TimerHistory start(Ticket task, LocalDateTime dateTime) throws AxelorException {
     Timer timer = find(task);
     return start(task, timer, dateTime);
   }
 
   @Override
-  public TimerHistory stop(TeamTask task, LocalDateTime dateTime) throws AxelorException {
+  public TimerHistory stop(Ticket task, LocalDateTime dateTime) throws AxelorException {
     Timer timer = find(task);
     return stop(task, timer, dateTime);
   }
 
   @Override
-  public void cancel(TeamTask task) {
+  public void cancel(Ticket task) {
     Timer timer = find(task);
     cancel(timer);
   }
 
   @Override
-  public Duration compute(TeamTask task) {
+  public Duration compute(Ticket task) {
     Duration total = Duration.ZERO;
     for (Timer timer : task.getTimerList()) {
       total = total.plus(compute(timer));